requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
gurobipy>=10.0
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65